+++ binary/src/IpcMessenger.ts
@@ -1,5 +1,5 @@
 import { IProtocol } from "core/protocol/index.js";
-import { IMessenger, type Message } from "core/util/messenger";
+import { IMessenger, type Message } from "core/protocol/messenger";
 import { ChildProcessWithoutNullStreams } from "node:child_process";
 import * as fs from "node:fs";
 import net from "node:net";
