+++ extensions/vscode/src/webviewProtocol.ts
@@ -4,7 +4,7 @@ import path from "path";
 import { FromWebviewProtocol, ToWebviewProtocol } from "core/protocol";
 import { WebviewMessengerResult } from "core/protocol/util";
 import { extractMinimalStackTraceInfo } from "core/util/extractMinimalStackTraceInfo";
-import { Message } from "core/util/messenger";
+import { Message } from "core/protocol/messenger";
 import { Telemetry } from "core/util/posthog";
 import { v4 as uuidv4 } from "uuid";
 import * as vscode from "vscode";
@@ -33,7 +33,8 @@ export async function showTutorial() {
 }
 
 export class VsCodeWebviewProtocol
-  implements IMessenger<FromWebviewProtocol, ToWebviewProtocol> {
+  implements IMessenger<FromWebviewProtocol, ToWebviewProtocol>
+{
   listeners = new Map<
     keyof FromWebviewProtocol,
     ((message: Message) => any)[]
@@ -127,7 +128,7 @@ export class VsCodeWebviewProtocol
             message = message.split("\n").filter((l: string) => l !== "")[1];
             try {
               message = JSON.parse(message).message;
-            } catch { }
+            } catch {}
             if (message.includes("exceeded")) {
               message +=
                 " To keep using Continue, you can set up a local model or use your own API key.";
@@ -181,7 +182,7 @@ export class VsCodeWebviewProtocol
     });
   }
 
-  constructor(private readonly reloadConfig: () => void) { }
+  constructor(private readonly reloadConfig: () => void) {}
   invoke<T extends keyof FromWebviewProtocol>(
     messageType: T,
     data: FromWebviewProtocol[T][0],
