+++ gui/src/context/IdeMessenger.ts
@@ -1,8 +1,8 @@
 import { ChatMessage, IDE, LLMFullCompletionOptions, PromptLog } from "core";
 import type { FromWebviewProtocol, ToWebviewProtocol } from "core/protocol";
 import { WebviewMessengerResult } from "core/protocol/util";
-import { MessageIde } from "core/util/messageIde";
-import { Message } from "core/util/messenger";
+import { MessageIde } from "core/protocol/messenger/messageIde";
+import { Message } from "core/protocol/messenger";
 import { createContext } from "react";
 import { v4 as uuidv4 } from "uuid";
 import "vscode-webview";
