+++ gui/src/hooks/useWebviewListener.ts
@@ -1,5 +1,5 @@
 import type { ToWebviewProtocol } from "core/protocol/index.js";
-import { Message } from "core/util/messenger";
+import { Message } from "core/protocol/messenger";
 import { useContext, useEffect } from "react";
 import { IdeMessengerContext } from "../context/IdeMessenger";
